Fix shape renaming bug
Until now, when renaming a shape in the list it changed the whole
object to the string with the new name instead of changing only the
name. This commit creates an attribute to shapes to hold their names and
changes the `setData` method in `DisplayFile` to take renaming into
account.