Fix shape orientation and transform combination
The `Rectangle` class creates objects with wrong reference points. That
is, the top-left point is actually at the bottom-left corner and the
botton-right is the top-right. Another problem is that the
`combine` function does not return the combined transform. Instead, it
returns the transform `t` used in the loop.

This commit fixes the creation of rectangles and, by consequence, the
window creation and its `normalize` function. It also fixes the
`combine` bug mentioned.